refactor: simplify webui projection tests
